chore: add PEP 561 type completeness for downstream mypy consumers
- Add -> None / -> PutResult return annotations to all public methods
  across components, metadata, and figure_registry submodules
- Cast pint .magnitude returns to float to satisfy warn_return_any
- Add __all__ to figure_registry/__init__.py covering all 65 public
  re-exports (previously only FlowSourceMode was explicitly re-exported)
- Enable no_implicit_reexport, warn_return_any, and warn_unused_ignores
  in [tool.mypy]; pre-commit mypy hook passes clean

+++ src/chemunited_core/connections/edge.py
@@ -225,12 +225,12 @@ class EdgeData(Element):
     @property
     def capacity(self) -> float:
         """capacity in SI"""
-        return self.length_value * np.pi * self.diameter**2 / 4  # m**3
+        return float(self.length_value * np.pi * self.diameter**2 / 4)  # m**3
 
     @property
     def length_value(self) -> float:
-        return self.length.to_base_units().magnitude
+        return float(self.length.to_base_units().magnitude)
 
     @property
     def diameter_value(self) -> float:
-        return self.diameter.to_base_units().magnitude
+        return float(self.diameter.to_base_units().magnitude)
